// src/components/Article/ColumnistBio/types.ts
export interface ContainerProps {
  bgColor: string;
  children: any
}

export interface ColumnistBioProps {
  authorTitle: string;
  authorTitleColor: string;
  bgColor: string;
  children: any;
  columnTitleColor: string;
  columnTitle: string;
  fontSizeAuthor: string;
  fontSizeColumn: string;
  fontWeightAuthor: string;
  fontWeightColumn: string;
}

export interface ContentProps {
  children: any
}

export interface ImageProps {
  children: any;
}

export interface ColumnTitleProps {
  children: any;
  color: string;
  fontSize: string;
  fontWeight: string
}

export interface AuhtorTitleProps {
  color: string;
  children: any;
  fontWeight: string
  fontSize: string
}

export interface TitlesProps {
  children: any
}